NoiseBackend: run reader thread at higher QoS
In macOS 13, Apple added priority inversion checks that trigger
because waiting on a Future blocks the waiting thread on the reader
thread.  Sometimes, though, it's convenient to block the main
thread with the understanding that that will block the UI if the
operation takes too long.  So this silences the warning for those
situations, although it doesn't feel like a great solution.