Racket: fix Val.fixnum for 0 when represented as NULL
Swift implicitly unwraps void * values and panics when it sees a NULL
pointer, but Chez represents a 0 fixnum as (void *)0, so Val.fixnum(0)
would blow up with:

    Unexpectedly found nil while implicitly unwrapping an Optional value

The fix is to make Val.ptr optional and annotate the C fixnum
procedures with _Nullable.  The other ptr-producing procedures don't
need the annotation since all other Chez pointers are guaranteed to be
non-zero, because fx 0 takes that spot.  However, if Chez ever changes
its fixnum representation, we'll have to adapt.

This would also be a problem for unboxed flonums, but the Racket C API
currently doesn't expose a way to interact with those, so we don't
need to do anything at the moment.